centeralize rootViewController initialization
