+++ GithubRepositoriesTask/AppDelegate.swift
@@ -16,21 +16,10 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
         // Override point for customization after application launch.
         window = UIWindow(frame: UIScreen.main.bounds)
-        let githubRepoService: GithubRepositoryServiceInterface = self.getRepositoryService(headers: nil)
-        let githubRepositoriesViewController: GithubRepositoriesViewController = GithubRepositoriesViewController(githubRepositoryService: githubRepoService)
-        let initialViewController: UINavigationController = UINavigationController(rootViewController: githubRepositoriesViewController)
-        window?.rootViewController = initialViewController
+        window?.rootViewController = RootViewController.initialized()
         window?.makeKeyAndVisible()
 
         return true
-    }
-
-    private func getRepositoryService(headers: [String: String]?) -> GithubRepositoryServiceInterface {
-        let requestHandler: RequestHandler = GithubFetcher(headers: headers)
-       let responseDecoder: ResponseDecoder = GithubRepositoriesDecoder()
-       let githubRepositoryRepo: GithubRepositoryReposioryInterface = GithubRepositoryRepository(requestHandler: requestHandler, responseDecoder: responseDecoder)
-
-        return GithubRepositoryService(githubRepositoryRepo: githubRepositoryRepo)
     }
 
     // MARK: UISceneSession Lifecycle
